fix: lua_print call is now protected; format with stylua
because sequins returns itself, if one called `require('sequins')` in
the REPL, the program would immediately crash, since sequins attempts
to turn itself into a string, which is problematic because the
sequins *class* has length zero. this fix protects against that in two
ways: one by altering our monkey-patched print function to use a
protected call so that the crash does not happen, and two by bailing
in sequins.__tostring if the `length` field of the sequins is nil.

closes #122