feat: remove notcurses, add libvaxis, minimal TUI mode
interesting under the hood changes:
- the event queue is now lock-free with a minimal amount of back-off
from the consumer when the queue is drained.
- event queue nodes are "intrusive"" and rely on `@fieldParentPtr` to
execute.
- as such, the event system is truly dynamic dispatch. while likely
slightly slower, this improves the workflow for adding a new module to
the system, and significantly improves the workflow for adding a new
event type.
- rather than passing a reader around to handle REPL events over
stdin, the CLI and TUI modes maintain a thread-safe queue of bytes
which the Lua VM peeks from. since event nodes are intrusive, new
readers from this queue do not need to be constructed.