license!: GPL -> MIT
with permission from the principal authors of norns, this commit relicenses seamstress versions 2.0.0 and above with the MIT license. since MIT is more permissive, especially in software that is composed with other software, it is my hope that this change will make using seamstress as an "engine" for more complete works an attractive future possibility.