feat!: export shared library for use with external lua processes
`zig build` now also compiles a shared library `seamstress.so` for use
with external Lua interpreters.

TODO: currently a user needs to provide the environment variable
SEAMSTRESS_LUA_PATH manually, but a best guess should be possible if
it is not available.

seamstress.quit() is removed in favor of
seamstress:stop(). seamstress:stop() attempts to clear the event loop,
but does not exit the program. new are seamstress:resume() and
seamstress:run(). the former runs the event loop for one tick, while
the latter takes optionally arguments fun, ..., which are fed to
seamstress.async.Promise. the latter also runs the event loop until it
is finished.

ptrFromHandle and handleFromPtr are now in lua_util, since that
pattern seems very common, as is a basic unrefCallback.

seamstress.cli and seamstress.repl are separated into distinct
modules and exposed to the module system. processChunk is removed from
lua_util in favor of calling seamstress.repl directly.

BREAKING CHANGE: Lua.allocator() panics when called on a Lua instance not
created by Zig, so all calls to Lua.allocator() should be replaced
with calls to the allocator() function in lua_util

BREAKING CHANGE: the seamstress configuration step is
removed. (running arbitrary Lua code feels less comfy lately, and most
use cases are moot with, e.g. the move to providing port numbers directly)